requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
googletrans==4.0.0rc1
redis==4.6.0
celery==5.3.1
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
python-dateutil==2.8.2
jsonschema==4.19.0
pydantic==2.8.0
openai==0.28.0
langdetect==1.0.9
geopy==2.3.0
